refactor game context into hooks
